+++ config/bootstrap.php
@@ -53,11 +53,26 @@ $defaults = [
 ];
 Configure::write(array_filter($defaults, fn (string $key): bool => !Configure::check($key), ARRAY_FILTER_USE_KEY));
 
-//Auto-discovers binaries
+/**
+ * It automatically discovers executables not already set by the user in the configuration.
+ *
+ * For `mysql` and `mysqldump` executables, it will first look for `mariadb` and `mariadb-dump` executables.
+ * It then normally searches all other possible executables canonically.
+ */
 $ExecutableFinder = new ExecutableFinder();
-$binaries = array_unique(array_merge(['bzip2', 'gzip'], ...array_values(array_map('array_values', DATABASE_BACKUP_EXECUTABLES))));
-foreach (array_filter($binaries, fn (string $binary): bool => !Configure::check('DatabaseBackup.binaries.' . $binary)) as $binary) {
-    Configure::write('DatabaseBackup.binaries.' . $binary, $ExecutableFinder->find($binary));
+foreach ([
+    'mariadb' => 'mysql',
+    'mariadb-dump' => 'mysqldump',
+] as $executable => $alias) {
+    if (!Configure::check('DatabaseBackup.binaries.'. $alias)) {
+        Configure::write('DatabaseBackup.binaries.' . $alias, $ExecutableFinder->find($executable));
+    }
+}
+$executables = array_merge(['bzip2', 'gzip'], ...array_values(array_map('array_values', DATABASE_BACKUP_EXECUTABLES)));
+foreach ($executables as $executable) {
+    if (!Configure::check('DatabaseBackup.binaries.' . $executable)) {
+        Configure::write('DatabaseBackup.binaries.' . $executable, $ExecutableFinder->find($executable));
+    }
 }
 
 //Checks for the target directory